### 2025-02-24 #7842 @editorialbot
[PRE REVIEW]: StochasticDominance.jl: A Julia Package for Higher Order Stochastic Dominance
**Submitting author:** @rajmadan96 (<a href="http://orcid.org/0009-0006-3273-9063">Rajmadan Lakshmanan</a>)
**Repository:** https://github.com/rajmadan96/StochasticDominance.jl.git
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.2
