+++ paper/paper.md
@@ -23,35 +23,24 @@ bibliography: "paper.bib"
 
 # Summary
 
-Stochastic dominance plays a key role in decision-making under uncertainty and quantitative finance. It compares random variables using their distribution functions. This concept helps to evaluate whether one investment, policy, or strategy is better than others in uncertain conditions. By using cumulative distribution functions, stochastic dominance allows decision-makers to make comparisons without assuming specific utility functions. It provides a mathematically rigorous method often used in optimization to maximize returns or minimize risk. Its precision and reliability make it a powerful tool for analyzing complex probabilistic systems.
+Stochastic dominance plays a key role in decision-making under uncertainty and quantitative finance. This concept helps to evaluate whether one investment, policy, or strategy is better than others in uncertain conditions. It provides a mathematically rigorous method often used in optimization to maximize returns or minimize risk. See `@Ogryczak:2002`, `@Dentcheva:2003`, `@Kuosmanen:2004`, `@Kopa:2017`, `@Kopa:2023`, `[@Maggioni:2016; @Maggioni:2019]`, and `@Consigli:2023` for extensive body of research on the applicability of stochastic dominance.
 
-Despite being a crucial tool in decision-making under uncertainty and quantitative finance, (higher-order) stochastic dominance involves infinitely many constraints, making it *computationally intractable* in practice. Our recent research `@Lakshmanan:2025` addresses this challenge by theoretically reducing the infinite constraints to a finite number. Building on this theoretical foundation, we have developed algorithms to maximize expected returns and minimize risk by satisfying the (higher-order) infinitely many stochastic dominance constraints. The paper aims to establish the mathematical completeness of this reduction and improve accessibility in research domains. However, no concrete, user-friendly implementation of (higher-order) stochastic dominance has been developed. Additionally, the existing prominent theoretical algorithms only discuss stochastic orders *two* and *three*, but not higher orders. Moreover, both, the discussion and implementation of non-integer orders, are absent.
-
-To address this gap, we present *StochasticDominance.jl*, an open-source Julia package tailored for verification and optimization under higher-order stochastic dominance constraints. Designed as a black-box solution, it enables users to input their data and obtain results effortlessly, without requiring extensive technical knowledge.
-
-There is an extensive body of research on the applicability of stochastic dominance across various domains. Numerous scholars, including `@Ogryczak:2002`, `@Dentcheva:2003`, `@Kuosmanen:2004`, `@Kopa:2017`, `@Kopa:2023`, `[@Maggioni:2016; @Maggioni:2019]`, and `@Consigli:2023`, have rigorously explored its significance.  
-Their work highlights the critical role of stochastic dominance in reducing computational complexity and establishing it as a powerful tool for solving complex optimization problems.  
-For a comprehensive recent textbook overview, see `@Dentcheva:2024`.
+Despite being a crucial tool, (higher-order) stochastic dominance involves infinitely many constraints, making it *computationally intractable* in practice. Our recent research `@Lakshmanan:2025` addresses this challenge by theoretically reducing the infinite constraints to a finite number. However, no concrete, user-friendly implementation of (higher-order) stochastic dominance has been developed. Additionally, the existing prominent theoretical algorithms only discuss stochastic orders *two* and *three*, but not higher orders. Moreover, both, the discussion and implementation of non-integer orders, are absent.
   
 # Main features of the package
 
[…]
