### 2025-03-11 #7842 @rajmadan96
Hi @samhforbes,

Thank you for the overview of the package.

We developed this package to facilitate the computational process of stochastic dominance. Until now, no concrete, user-friendly implementation of (higher-order) stochastic dominance has been available. Moreover, existing prominent theoretical algorithms primarily discuss stochastic orders two and three but do not address higher orders. Additionally, discussions and implementations of non-integer orders are entirely absent.

To address these gaps, we designed this package as a black-box solution that allows users to input their data and obtain results effortlessly, without requiring extensive technical knowledge. While offering these advantages, the package is grounded in significant mathematical understanding, making it both concise and powerful.

Given its merits, I believe the package has the potential to be reviewed in JOSS. I have reduced the paper to 991 words accordingly.

Sincerely,
Rajmadan Lakshmanan
