### 2025-03-21 #7842 @lrnv
@FlyingWorkshop, @JanJereczek, @AlCap23, Would any of you be motivated and feel pertinent to review this submission for JOSS ?
