### 2025-03-27 #7842 @lrnv
@ViktorAJStein, @felixbartel, @bfpc, @odow greetings ! Would any of you be motivated and feel pertinent to review this submission for JOSS ? If you do not already know about JOSS please tell me I'll explain what this is.
