### 2025-03-27 #7842 @lrnv
@ViktorAJStein If on the other hand you have knowledge about the subject of the package, your input might still be relevant, the fact tht you do not know Julia that much is not a deal breaker for me. Tell me what you think.
