### 2025-03-27 #7842 @ViktorAJStein
@lrnv I briefly looked over the paper and I am not familiar with the key concepts, for example I have not heard of stochastic dominance before today.
