### 2025-04-28 #7842 @kellino
Dear Irnv,

Sorry to say that I’ve got no available free time until at least August, so I won’t be able to review this paper. Sorry again.

Kind regards,

David

> On 28 Apr 2025, at 08:02, Oskar Laverny ***@***.***> wrote:
>
>
> lrnv
>  left a comment
> (openjournals/joss-reviews#7842)
>  <https://github.com/openjournals/joss-reviews/issues/7842#issuecomment-2834175710>
> @kellino <https://github.com/kellino>, @niyiyu <https://github.com/niyiyu>, @nluetts <https://github.com/nluetts> would any of you be available to review this submission for JOSS ?
>
