### 2025-05-21 #7842 @lrnv
@Guganesan-Ilavarasan Absolutely, let me give you a few details.

JOSS publishes articles about open source research software. It is a free, open-source, community driven and developer-friendly online journal, with a unique review process.

The review process takes place in a GitHub issue (that I will soon open), is open, and author-reviewer-editor conversations are encouraged.

JOSS reviews involve downloading and installing the software, and inspecting the repository and submitted paper for key elements. See https://joss.readthedocs.io/en/latest/review_criteria.html for more details on the criteria. The process is checklist-driven to ensure consistency between the reviews.

Editors and reviewers post comments on the Review issue, and authors respond to the comments and improve their submission until acceptance (or withdrawal, if they feel unable to satisfy the review).

I will now close this thread and open the review thread so that we can discuss it there if you want.
