### 2025-05-26 cf9a338
[paper] fix references and tidy Julia code examples
