### 2025-05-26 rajmadan96/StochasticDominance.jl#8 @odow
JOSS paper changes
x-ref https://github.com/openjournals/joss-reviews/issues/8255

> A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

See, e.g., the example paper: https://joss.readthedocs.io/en/latest/example_paper.html

> Community guidelines: Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support

This could just be a paragraph at the bottom of https://github.com/rajmadan96/StochasticDominance.jl/blob/main/docs/src/index.md

> Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?

The paper is generally well written, but you should review it for grammar. (As one example, "See [tutorials] for comprehensive technical explanation.")
