### 2025-05-26 #8255 @Guganesan-Ilavarasan
Hi @Irnv,

I've completed my review of the [```StochasticDominance.jl```](https://github.com/rajmadan96/StochasticDominance.jl) Julia package and recommend acceptance with some suggestions for improvement. The package is well-designed and documented, making it easy for practitioners to use higher-order stochastic dominance tools with clear documentation and tutorials. The software installs and runs as illustrated. Further, the accompanying paper is well-written and aligns with JOSS's scope; the repository includes an OSI-approved license.

**Strengths**

- Well-written tutorials and documentation.

- Functions and their documentation are user-friendly with sufficient examples.

- The paper is clear and adequately references the field.

**Recommendations**

- A separate "Statement of Need" section is missing. It is integrated into the summary, but a separate section would be appreciated.

-  In the tutorials, I recommend removing the ```julia>``` prompt in the example and tutorial codes for a cleaner presentation.

-  I like the message prompt from the ```verify_dominance``` function; to express the optimal allocation and optimal values clearly, I recommend including similar prompts for the other two functions.

**Typos**

- ```println("Optimal Objective: expected_portfolio_return(x_opt,ξ,p_ξ)", )``` should be ```println("Optimal Objective:", expected_portfolio_return(x_opt,ξ,p_ξ))```.

Finally, the ```RiskFunction``` is not correctly integrated into the main module. Please fix this.

Thanks! It was a nice experience to learn and review the ```StochasticDominance.jl``` package. I wish the authors all the best!

Fondly, Gugan
