### 2025-06-04 291242f
Update paper.md
- Added statement of interest 
- Enhanced the writing (grammars )
