### 2025-06-04 0dd4c11
Update index.md
added community guidelines
