+++ docs/src/index.md
@@ -22,10 +22,10 @@ The `StochasticDominance.jl` package provides tools to:
 
 The package `StochasticDominance.jl` can be installed in Julia REPL as follows:
 
-```julia
-using Pkg
-Pkg.add("StochasticDominance")
-using StochasticDominance
+```julia-repl
+julia> using Pkg
+julia> Pkg.add("StochasticDominance")
+julia> using StochasticDominance
 ```
 
 Once you have installed `StochasticDominance.jl`, we recommend going through the tutorials from beginning to end to understand how to use the package to verify stochastic dominance and determine the optimal allocation between a benchmark asset and a portfolio.
