### 2025-06-04 fbdf533
Update tutorial3.md
