+++ docs/src/tutorial/tutorial4.md
@@ -45,37 +45,37 @@ Here, $\xi_0 \in \mathbb{R}^{n}$ represents the return of a benchmark portfolio,
 
 In the first example, we illustrate the function `optimize_min_riskreturn_SD` using representative data.
 
-```julia-repl
+```julia
 # Define portfolio return scenarios (rows: assets, columns: scenarios)
-julia> ξ = [0.02  0.05 -0.01;  # Asset 1 returns under different scenarios
+ξ = [0.02  0.05 -0.01;  # Asset 1 returns under different scenarios
      0.03  0.06  0.02]   # Asset 2 returns under different scenarios
 
 # Define benchmark return scenarios (column vector)
-julia> ξ_0 = [0.01, 0.04, 0.00]  # Benchmark returns in the same scenarios
+ξ_0 = [0.01, 0.04, 0.00]  # Benchmark returns in the same scenarios
 
 # Define probability distributions for portfolio and benchmark scenarios
-julia> p_ξ = [0.3, 0.4, 0.3]     # Probabilities for each scenario
-julia> p_ξ_0 = [0.4, 0.4, 0.2]   # Probabilities for benchmark scenarios
+p_ξ = [0.3, 0.4, 0.3]     # Probabilities for each scenario
+p_ξ_0 = [0.4, 0.4, 0.2]   # Probabilities for benchmark scenarios
 
 # Set the order of stochastic dominance
-julia> SDorder = 2  # Second-order stochastic dominance
+SDorder = 2  # Second-order stochastic dominance
 
 # Set the risk parameters
 β=0.5
 r=2.0
 
-julia> x_opt, q_opt,t_opt = optimize_min_riskreturn_SD(ξ, ξ_0,SDorder;p_ξ, p_ξ_0,β,r,verbose=true)
+x_opt, q_opt,t_opt = optimize_min_riskreturn_SD(ξ, ξ_0,SDorder;p_ξ, p_ξ_0,β,r,verbose=true)
 ```
 ### Output
 
-```julia-repl
+```julia
 Simplex Constraints residuals: 5.903068123425115e-6
 Stochastic Dominance Constraints residuals: 0.0"
-julia> println("Optimal asset allocation x: ",x_opt)
+println("Optimal asset allocation x: ",x_opt)
 Optimal asset allocation x: [0.004074767447193794, 0.9959311356209297]
-julia> println("Optimal q: ",q_opt) 
+println("Optimal q: ",q_opt) 
 Optimal q: -0.019892461315708196
-julia> println("Optimal Objective: ", riskfunction_asset_allocation(x_opt, q_opt, ξ, r, p_ξ, β))
+println("Optimal Objective: ", riskfunction_asset_allocation(x_opt, q_opt, ξ, r, p_ξ, β))
 Optimal Objective: -0.019876482848988134
 ```
 In general, the output provides the optimal asset allocation vector $x$ of length $d$, the optimal $q$ for a higher-order risk measure, and a single optimal $t$ that confirms the stochastic dominance of the given order. For more technical details on reducing the verification of infinitely many $t$ (i.e., $t \in \mathbb{R}$) to a single value, refer to [Lakshmanan et al. 2025](https://arxiv.org/abs/2501.14565).
[…]
