### 2025-06-06 #8255 @rajmadan96
Dear @odow,

Thank you for the detailed review and recommendations. Below, we have addressed your recommendations and outlined the changes we made:

_The paper is generally well written, but you should review it for grammar.
(Example: “See [tutorials] for comprehensive technical explanation.”)_
**Response:** Thank you for your appreciation. We have carefully reviewed the manuscript and corrected the grammatical issues, including the one you highlighted.

_Statement of Need_
**Response:** As suggested, we have added a dedicated “Statement of Need” section in the updated version of the paper.

_Community guidelines: Are there clear guidelines for third parties wishing to (1) contribute to the software, (2) report issues or problems, (3) seek support?_
**Response:** Thank you for the suggestion. We have now included a "Contributing and Support" section on the repository's main page, providing clear instructions for contributions, issue reporting, and support.

_verify_dominance claims to return a message, but it only prints to the screen and returns nothing. Consider returning a Bool instead._
**Response:** Indeed, you are right! We’ve updated the verify_dominance function so that it now returns a Bool indicating whether dominance holds.

Thanks for fixing the references and updating Julia code from the old syntax.

Your valuable feedback has improved the clarity and usability of the package. Thank you once again for your contribution.

Best regards,
Rajmadan Lakshmanan and Alois Pichler
