### 2025-06-06 rajmadan96/StochasticDominance.jl#8 @rajmadan96
Thank you for your thoughtful feedback and appreciation. We have carefully reviewed the manuscript and addressed the grammatical issues, including the one you highlighted.

As suggested, we have added a dedicated “Statement of Need” section to the updated version of the paper. Additionally, we have included a “Contributing and Support” section on the repository’s main page, offering clear guidelines for contributions, issue reporting, and seeking support.
