### 2025-06-09 04d76b6
Update paper.bib
updated DOI
