### 2025-06-09 b690dbc
Update paper.bib
