+++ paper/paper.bib
@@ -62,7 +62,7 @@ doi     = {10.48550/arXiv.2501.14565},
   year      = {2023},
   pages     = {1--21},
   publisher = {Springer},
-  url       = {10.1007/s10479-023-05658-y},
+  doi       = {10.1007/s10479-023-05658-y},
 }
 
 
