### 2025-06-10 rajmadan96/StochasticDominance.jl#9 @lrnv
Docs improvement
I took a glimpse into the docs, and reading this page https://rajmadan96.github.io/StochasticDominance.jl/dev/tutorial/tutorial1/ to discover the topic left me somewhat unsatisfied:

- The tone and structure of the document is not as formal as it could be.
- You use the "stochastic docminance of order p" in your argument before you define it correctly. Not per-se an issue but it makes your argument harder to read...
- What is `essinf` ?
- What is the stochastic order with respect to the p-norm ?
- Can you add references to the litterature ? e.g. using [DocumenterCitations.jl](https://github.com/JuliaDocs/DocumenterCitations.jl) or at least pointers to the right wikipedia pages if there are some ?

Moreover, the set $\Chi$ is weird... is can contain several different variables, "often only one" as you say, but the definition of the stochastic order does not precise which ones ? So how can i know ?

--------

More precisely, I do not expect your documentation to be a textbook-level lecture on the subject, but at least it should be tidy and easy to understand. I will continue reading the rest but i expect the issue to be all-over the theoretical parts so maybe you could check for these things everywhere in the docs while proof-reading?
