### 2025-06-10 #8255 @lrnv
Keeping track: https://github.com/rajmadan96/StochasticDominance.jl/issues/9
