+++ paper/paper.bib
@@ -97,6 +97,18 @@ doi     = {10.48550/arXiv.2501.14565},
   url    = {https://optimization-online.org/wp-content/uploads/2023/11/ConsigliDentchevaMaggioniMicheli-1.pdf},
 }
 
+@Article{Pichler:2024,
+  title={Connection between higher order measures of risk and stochastic dominance},
+  author={Pichler, Alois},
+  journal={Computational Management Science},
+  volume={21},
+  number={2},
+  pages={41},
+  year={2024},
+  doi={10.1007/s10287-024-00523-0}	
+  publisher={Springer}
+}
+
 
 @Book{Dentcheva:2024,
   title     = {Risk-Averse Optimization and Control: Theory and Methods},
