### 2025-06-10 rajmadan96/StochasticDominance.jl#9 @odow
Not to over step here, but:

> The tone and structure of the document is not as formal as it could be.

Surely this is not a requirement for JOSS, or for a Julia package. It's an implementation of an algorithm, not a math paper.
