+++ docs/src/tutorial/tutorial1.md
@@ -27,7 +27,7 @@ From a portfolio optimization perspective, the feasible random variable $Y$ in t
 
 ## Optimal Portfolio Scaling under Stochastic Dominance of Order $p$
 
-Given a **fixed benchmark** random variable $ \xi_0 \in L^p$ and a **portfolio return** random variable \( \xi \in L^p \), we aim to determine an **optimal scaling factor** $x > 0$ such that the scaled portfolio $x^{\top}\xi$ dominates $ \xi_0 $ under **stochastic dominance of order \( p \)**.
+Given a **fixed benchmark** random variable $ \xi_0 \in L^p $ and a **portfolio return** random variable $ \xi \in L^p $, we aim to determine an **optimal scaling factor** $x > 0$ such that the scaled portfolio $x^{\top}\xi$ dominates $ \xi_0 $ under stochastic dominance of order.
 
 ### Goal:
 Find \( x > 0 \) such that:
