+++ docs/src/tutorial/tutorial1.md
@@ -27,10 +27,10 @@ From a portfolio optimization perspective, the feasible random variable $Y$ in t
 
 ## Optimal Portfolio Scaling under Stochastic Dominance of Order $p$
 
-Given a **fixed benchmark** random variable $ \xi_0 \in L^p $ and a **portfolio return** random variable $ \xi \in L^p $, we aim to determine an **optimal scaling factor** $x > 0$ such that the scaled portfolio $x^{\top}\xi$ dominates $ \xi_0 $ under stochastic dominance of order.
+Given a **fixed benchmark** random variable $\xi_0 \in L^p$ and a **portfolio return** random variable $\xi \in L^p$, we aim to determine an **optimal scaling factor** $x > 0$ such that the scaled portfolio $x^{\top}\xi$ dominates $ \xi_0 $ under stochastic dominance of order.
 
 ### Goal:
-Find \( x > 0 \) such that:
+Find $x > 0$ such that:
 
 ```math
 \xi_0 \preccurlyeq^{(p)} x^{\top}\xi \quad \text{or} \quad \mathbb{E}(t - \xi_0)_+^{p-1} \geq \mathbb{E}(t - x^{\top}\xi )_+^{p-1}, \quad \forall t \in \mathbb{R}.
