+++ docs/src/tutorial/tutorial1.md
@@ -56,7 +56,7 @@ Find $x > 0$ such that:
 This approach is especially useful in **portfolio optimization** where a benchmark (e.g., market index) is fixed, and the investor seeks to optimally scale their portfolio to satisfy risk-return tradeoffs under rigorous dominance conditions.
 
 
-For other equivalent formulations and technical details, we refer to [Dentcheva and Ruszczyński (2024)]([https://doi.org/10.1007/s10589-015-9758-0](https://doi.org/10.1007/978-3-031-57988-2)).
+For other equivalent formulations and technical details, we refer to [Dentcheva and Ruszczyński (2024)](https://doi.org/10.1007/s10589-015-9758-0).
 
  
 This package provides tools to analyze stochastic dominance, enabling users to apply these concepts effectively in decision-making and financial optimization.
