+++ docs/Project.toml
@@ -1,4 +1,5 @@
 [deps]
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
 StochasticDominance = "8041b248-d9a2-49d2-9c7f-91ddf745d052"
+DocumenterCitations = "daee34ce-89f3-4625-b898-19384cb65244"
 
