+++ docs/src/refs.bib
@@ -1,0 +1,53 @@
+
+@Book{Dentcheva:2024,
+  title     = {Risk-Averse Optimization and Control: Theory and Methods},
+  publisher = {Springer Nature},
+  year      = {2024},
+  author    = {Dentcheva, Darinka and Ruszczy{\'n}ski, Andrzej},
+  doi       = {10.1007/978-3-031-57988-2},
+}
+
+
+@Article{Lakshmanan:2025,
+author  = {Lakshmanan, Rajmadan and Pichler, Alois and Kopa, Milo{\v{s}}},
+title   = {Higher-Order Stochastic Dominance Constraints in Optimization},
+journal = {arXiv preprint arXiv:2501.14565},
+year    = {2025},
+doi     = {10.48550/arXiv.2501.14565},
+}
+
+@Article{Dentcheva:2003,
+  author    = {Dentcheva, Darinka and Ruszczy{\'n}ski, Andrzej},
+  title     = {Optimization with stochastic dominance constraints},
+  journal   = {SIAM Journal on Optimization},
+  year      = {2003},
+  volume    = {14},
+  number    = {2},
+  pages     = {548--566},
+  doi       = {10.1137/S1052623402420528},
+  publisher = {SIAM},
+}
+
+@Article{Kopa:2017,
+  author  = {Post, Thierry and Kopa, Milo\v{s}},
+  title   = {Portfolio Choice Based on Third-Degree Stochastic Dominance},
+  journal = {Management Science},
+  year    = {2017},
+  volume  = {63},
+  number  = {10},
+  pages   = {3381-3392},
+  doi     = {10.1287/mnsc.2016.2506},
+}
+
+@Article{Pichler:2024,
+  title={Connection between higher order measures of risk and stochastic dominance},
+  author={Pichler, Alois},
+  journal={Computational Management Science},
+  volume={21},
+  number={2},
+  pages={41},
+  year={2024},
+  doi={10.1007/s10287-024-00523-0}	
+  publisher={Springer}
+}
+
