+++ docs/src/refs.bib
@@ -1,5 +1,5 @@
 
-@Book{Dentcheva:2024,
+@book{Dentcheva2024,
   title     = {Risk-Averse Optimization and Control: Theory and Methods},
   publisher = {Springer Nature},
   year      = {2024},
@@ -7,8 +7,7 @@
   doi       = {10.1007/978-3-031-57988-2},
 }
 
-
-@Article{Lakshmanan:2025,
+@article{Lakshmanan2025,
 author  = {Lakshmanan, Rajmadan and Pichler, Alois and Kopa, Milo{\v{s}}},
 title   = {Higher-Order Stochastic Dominance Constraints in Optimization},
 journal = {arXiv preprint arXiv:2501.14565},
@@ -16,7 +15,7 @@ year    = {2025},
 doi     = {10.48550/arXiv.2501.14565},
 }
 
-@Article{Dentcheva:2003,
+@article{Dentcheva2003,
   author    = {Dentcheva, Darinka and Ruszczy{\'n}ski, Andrzej},
   title     = {Optimization with stochastic dominance constraints},
   journal   = {SIAM Journal on Optimization},
@@ -28,7 +27,7 @@ doi     = {10.48550/arXiv.2501.14565},
   publisher = {SIAM},
 }
 
-@Article{Kopa:2017,
+@article{Kopa2017,
   author  = {Post, Thierry and Kopa, Milo\v{s}},
   title   = {Portfolio Choice Based on Third-Degree Stochastic Dominance},
   journal = {Management Science},
@@ -39,7 +38,7 @@ doi     = {10.48550/arXiv.2501.14565},
   doi     = {10.1287/mnsc.2016.2506},
 }
 
-@Article{Pichler:2024,
+@article{Pichler2024,
   title={Connection between higher order measures of risk and stochastic dominance},
   author={Pichler, Alois},
   journal={Computational Management Science},
