+++ docs/src/tutorial/tutorial1.md
@@ -57,7 +57,7 @@ Find $x > 0$ such that:
 ### Practical Use:
 This approach is especially useful in **portfolio optimization** where a benchmark (e.g., market index) is fixed, and the investor seeks to optimally scale their portfolio to satisfy risk-return tradeoffs under rigorous dominance conditions.
 
-This package provides tools to analyze stochastic dominance, enabling users to apply these concepts effectively in decision-making and financial optimization. The mathematical background of this package was discussed in [Lakshmanan:2025](@citet).
+This package provides tools to analyze stochastic dominance, enabling users to apply these concepts effectively in decision-making and financial optimization. The mathematical background of this package was discussed in [Lakshmanan2025](@citet).
 
 # References
 
