### 2025-06-11 rajmadan96/StochasticDominance.jl#9 @rajmadan96
Thanks for the comments, @Irnv! To cater to a broader audience, we’ve removed some of the auxiliary theoretical explanations and added clearer, more straightforward descriptions of the concepts. We’ve also included a reference to the stochastic dominance Wikipedia page and added sufficient academic references (DocumenterCitations.jl) to help guide users in applying the package.

@odow Thanks for stepping in — you’re right. However, I’ve taken the editor’s suggestions into account to improve clarity and broaden the package’s accessibility, which might enhance its reach.
