+++ paper/paper.bib
@@ -105,7 +105,7 @@ doi     = {10.48550/arXiv.2501.14565},
   number={2},
   pages={41},
   year={2024},
-  doi={10.1007/s10287-024-00523-0}	
+  doi={10.1007/s10287-024-00523-0},	
   publisher={Springer}
 }
 
