### 2025-06-11 #8255 @rajmadan96
@lrnv  I finished the author tasks. I released a new version, v1.0.4, which is archived at Zenodo: [https://doi.org/10.5281/zenodo.15588305](https://doi.org/10.5281/zenodo.15642573) .
