### 2025-06-11 #8255 @Guganesan-Ilavarasan
> [@Guganesan-Ilavarasan](https://github.com/Guganesan-Ilavarasan) Your oppinon ?

Hi, @Irnv!

Apologies for the late receipt; the revised corrections are **_satisfactory_** to my raised remarks.

Many congratulations to both authors!

Godspeed, [Gugan](https://www.linkedin.com/in/guganesan-ilavarasan-b5571b16b/)
