### 2025-06-11 #8255 @rajmadan96
> [@rajmadan96](https://github.com/rajmadan96) The title of the doi archive should match the paper's title, can you modify the archive title ?

@lrnv Apologies! I changed it, it matches now
