+++ paper/paper.bib
@@ -93,8 +93,10 @@ doi     = {10.48550/arXiv.2501.14565},
 @Article{Consigli:2023,
   author = {Consigli, Giorgio and Dentcheva, Darinka and Maggioni, Francesca and Micheli, Giovanni},
   title  = {Asset liability management under sequential stochastic dominance constraints},
+journal={Optimization Online}
   year   = {2023},
-  url    = {https://optimization-online.org/wp-content/uploads/2023/11/ConsigliDentchevaMaggioniMicheli-1.pdf},
+doi={10.48550/arXiv.2505.16486}
+  ,
 }
 
 @Article{Pichler:2024,
@@ -127,28 +129,3 @@ doi     = {10.48550/arXiv.2501.14565},
   number  = {23-22},
   doi     = {10.2139/ssrn.4629613},
 }
-
-
-
-@Manual{JULIA2019,
-	author={Julia},
-	year = {2019},
-	address={MIT},
-	organization={Julia Computing},
-	title={{Julia v1.2}},
-	url={https://docs.julialang.org/en/v1}
-}
-
-
-@article{julia,
-    title={Julia: A fresh approach to numerical computing},
-    author={Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B},
-    journal={SIAM {R}eview},
-    volume={59},
-    number={1},
-    pages={65--98},
-    year={2017},
-    publisher={SIAM},
-    doi={10.1137/141000671}
-}
-
