### 2025-06-11 1d94abb
Update paper.bib
