+++ paper/paper.bib
@@ -93,10 +93,9 @@ doi     = {10.48550/arXiv.2501.14565},
 @Article{Consigli:2023,
   author = {Consigli, Giorgio and Dentcheva, Darinka and Maggioni, Francesca and Micheli, Giovanni},
   title  = {Asset liability management under sequential stochastic dominance constraints},
-journal={Optimization Online}
+  journal={Optimization Online}
   year   = {2023},
-doi={10.48550/arXiv.2505.16486}
-  ,
+  doi={10.48550/arXiv.2505.16486},
 }
 
 @Article{Pichler:2024,
