### 2025-06-11 be32287
Update paper.bib
