### 2025-06-11 #8255 @rajmadan96
> [@rajmadan96](https://github.com/rajmadan96) Moreover, two references are doi-less can you check them up and see if something more precise can be referenced ? e.g. Julia 1.2 looks like a very weird ref.

@lrnv Thanks, fixed
