+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"api/#API-Reference","page":"API Reference","title":"API Reference","text":"","category":"section"},{"location":"api/#Main-Functions","page":"API Reference","title":"Main Functions","text":"","category":"section"},{"location":"api/","page":"API Reference","title":"API Reference","text":"```@docs StochasticDominanceMeanReturn StochasticDominanceRiskMeasure AuxiliaryFunc.safeexponent AuxiliaryFunc.simplex1 AuxiliaryFunc.simplex2 AuxiliaryFunc.gpminus1 AuxiliaryFunc.gp AuxiliaryFunc.gbar AuxiliaryFunc.boundary_moments AuxiliaryFunc.CheckConvergenceSimplex AuxiliaryFunc.CheckConvergenceSD AuxiliaryFunc.MeanReturn AuxiliaryFunc.plotOptimalAssetAllocationMeanReturn AuxiliaryFunc.RiskFunction AuxiliaryFunc.BenchmarkRiskFunction AuxiliaryFunc.plotOptimalAssetAllocationRiskFunction Newton PSO ","category":"page"},{"location":"tutorial/tutorial1/","page":"Introduction","title":"Introduction","text":"CurrentModule = StochasticDominance","category":"page"},{"location":"tutorial/tutorial1/#Introduction","page":"Introduction","title":"Introduction","text":"","category":"section"},{"location":"tutorial/tutorial1/","page":"Introduction","title":"Introduction","text":"A stochastic program is a mathematical program that involves some uncertain data. These parameters may be mostly accurately described by random variables. In most cases, it is difficult to optimize directly in terms of the distributions of these random variables. Hence, in most cases, these distributions are approximated by discrete distributions with a finite number of scenarios for the random variables. This discretization procedure is what is often called scenario generation. Uncertainty in long-term capacity planning is inescapable. The random parameters can be conceived to follow a multistage stochastic process over some time space so that the discrete scenarios represent sample paths. The approach we take is to form an approximation of the original stochastic process by discretization.","category":"page"},{"location":"tutorial/tutorial1/","page":"Introduction","title":"Introduction","text":"In multistage stochastic optimization, we are interested in approximations of stochastic processes by finite structures. These processes are random and they have uncertain scenarios and a decision maker needs to make decisions at different stages of the process. It is useful to depict the possible sequences of data for this processes in form of a scenario tree in the case of a discrete time stochastic process and a scenario lattice for Markovian data processes.","category":"page"},{"location":"tutorial/tutorial1/","page":"Introduction","title":"Introduction","text":"A scenario tree is a set of nodes and branches used in models of decision making under uncertainty. Every node in the tree represents a possible state of the world at a particular point in time and a position where a decision can be made. Each tree node has a single predecessor and multiple successors whereas a lattice can have many predecessor.","category":"page"},{"location":"tutorial/tutorial1/","page":"Introduction","title":"Introduction","text":"A scenario tree/lattice is organized in levels which corresponds to stages 1ldotsT. Each node in a stage has a specified number of predecessors as defined by the branching structure. A node represents a possible state of the stochastic process and the vertices represents the possibility of transition between the two connected nodes. A scenario tree differs from a scenario lattice by the condition that each node in stage t must have one predecessor in stage t-1. For a lattice, that is not the case; all the nodes in stage t-1 share the same children in stage t.","category":"page"},{"location":"tutorial/tutorial1/#Goal-of-StochasticDominance.jl","page":"Introduction","title":"Goal of StochasticDominance.jl","text":"","category":"section"},{"location":"tutorial/tutorial1/","page":"Introduction","title":"Introduction","text":"We model stochastic pro
[…]
