### 2025-06-12 #8255 @lrnv
> @editorialbot accept

I agree the sentence is misleading, it is just addressed to editors in chief not to authors ;) pleas simply wait for the track EiC to take care of the paperwork from now on
