### 2025-06-20 #8255 @samhforbes
Hi @rajmadan96 this is looking good, the only thing I wondered was whether you wanted the citations in the first paragraph to render as only having the years in parenthesis? See [https://pandoc.org/MANUAL.html#citation-syntax](https://pandoc.org/MANUAL.html#citation-syntax)
