### 2025-06-20 #8255 @rajmadan96
Hi @samhforbes, thanks! We prefer to include the name and year (if possible), as it immediately guides the reader to explore other works by the same author—especially those closely related to Stochastic Dominance. For instance, Dentcheva & Ruszczyński are pioneers in this field, and referencing them in this way could be particularly helpful for readers.
