### 2025-06-20 #8255 @samhforbes
Sorry @rajmadan96 I probably wasn't clear - I mean to have Dentcheva & Ruszczyński (Year) rather than (Dentcheva & Ruszczyński , Year) in that paragraph
