### 2025-06-20 #8255 @rajmadan96
@samhforbes, why does the page say 404 error? Any idea?

 https://doi.org/10.21105/joss.08255
